+++ MagicAPI/src/main/java/com/elmakers/mine/bukkit/api/magic/Mage.java
@@ -598,6 +598,7 @@ public interface Mage extends CostReducer, CooldownReducer {
     boolean isAutomaton();
 
     Double getAttribute(String attributeKey);
+    void setAttribute(String attributeKey, Double attributeValue);
 
     /**
      *